mavlink: move GIMBAL_DEVICE_SET_ATTITUDE to separate stream header
